+++ src/java/chat/voiceup/openfire/plugin/httpfileupload/Component.java
@@ -15,7 +15,7 @@
  *
  */
 
-package org.igniterealtime.openfire.plugins.httpfileupload;
+package chat.voiceup.openfire.plugin.httpfileupload;
 
 import org.apache.maven.model.Model;
 import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
