+++ src/java/chat/voiceup/openfire/plugin/httpfileupload/LambdaSlotService.java
@@ -1,11 +1,9 @@
-package org.igniterealtime.openfire.plugins.httpfileupload;
+package chat.voiceup.openfire.plugin.httpfileupload;
 
 import com.amazonaws.services.lambda.*;
 import com.amazonaws.services.lambda.model.InvokeRequest;
 import com.amazonaws.services.lambda.model.InvokeResult;
 import com.google.gson.Gson;
-
-import java.nio.charset.Charset;
 
 public class LambdaSlotService implements SlotService {
