+++ src/java/chat/voiceup/openfire/plugin/httpfileupload/VoiceSlotService.java
@@ -1,4 +1,4 @@
-package org.igniterealtime.openfire.plugins.httpfileupload;
+package chat.voiceup.openfire.plugin.httpfileupload;
 
 import com.google.gson.Gson;
 import org.jivesoftware.util.Log;
